Try to use tilde version of url if dot version fails
mongo 2.6+ doesnt support dots in key names so when saving
they are now saved with tilde's.  however, some older installs
still have the dots so I try to manipulate the dot version first
and if that fails then I try the tilde version.  If that fails
I leave how the system already handled that scenario.