Generate a secure token for invites.
Unless you're planning to do validation on the token based on it's construction (e.g. comparing a password against a hashed and salted password), there is no benefit to artificially reducing entropy by feeding deterministic and computable into the hash.

In this instance what you're doing is feeding time() (easily determined by the attacker) + email (supplied by the attacker) + a random value with highly limited scope. Additionally, where email isn't there, you hash a hash, which adds no extra entropy.

True, an invitation must exist, so it's not actually all that of a problem here, but 1) we have a token, so use it. 2) it's a smaller function, 3) perhaps most importantly, we should get out of the habit of reducing entropy when we don't need to.