set CURL_SAFE_UPLOAD to TRUE (PHP 5.5+)
this prevents curl from interpreting values with a leading @ as
filenames, and is the default value in PHP 5.6.

otherwise, when sending e.g. Tweets that start with an @ via
Webservice, curl tries to interpret it as a filename and chokes.

this does not help PHP 5.4, where it does not seem to be possible
to send textual values that start with an @.