remove ActivityStreamPost
as discussed on the mailing list (https://groups.google.com/forum/#!topic/known-dev/7oAOmiwjYaA),
ActivityStreamPost tends to get in the way, as the data model has evolved
away from a pure AS 1.0 representation.

the big win in eliminating it is seriously cutting down on the number
of SQL queries required to render a feed.