Make micropub authentication work on FastCGI servers
When Apache connects to PHP via FastCGI, the HTTP "Authorization" header
is not sent to PHP.
.htaccess is modified so that the header is made available to PHP.

The header is available as REDIRECT_HTTP_AUTHORIZATION, so the
Idno/Common/Page::getallheaders() is adjusted to take that into account.
Usage of the native `getallheaders()` function is dropped since it does not
return the REDIRECT_HTTP_AUTHORIZATION header at all.

Resolves: #1505, #1509