Added syndication plugin for pnut social network
pnut is a Twitter-like status update network
Added note to plugin for App.net, which has been shut down. Please keep the link to the plugin as it was actually used as a template for the pnut plugin and may be useful to others in the future for this reason.