Avoid potential CSRF exploit by binding token to current session ID.
Using the refresh token endpoint, it is theoretically possible for an evil site to use a backend process to retrieve
a valid token for a given URL by faking XHR, e.g.

```curl -H "X-Requested-With: XMLHttpRequest" -I http://good.known.com/```

They could then present a form with the valid token which could be used to trigger the form.

Binding the token to the current session means that, absent a session fixation attack (which is a bigger issue), the backend process will return an entirely different token to the one presented by the user's browser.