Use Translation objects rather than passing low level strings
Previously, while you registered a translation using a Translation object, Language still registered strings. This was inflexible, as it locked translations into the massive array of strings approach and made it hard to replace with something else (e.g. gettext).

The solution was to store the Translation objects themselves, so that it didn't matter how the translation object was actually stored.