Diagnostics: Adding a diagnostics warning for multiple configs
Under certain circumstances it seems that some domains saved multiple configuration entries. This caused issues with token generation because of a secondary issue with SQL data handlers (fixed in c34567f). 

In order to get a better idea on where this occurs, I've added a diagnostic check for this case.