Catch fatal errors in object instantiation
The switch to having the Entity abstract caused issues where the database contained objects created by failed unit tests. When these now abstract objects were instantiated, a fatal error was thrown.

PHP 7 allows for catching of fatal errors, so the solution was to catch and log these errors.

Requires PHP7+ to work, 5.x should still run, but the error will not be caught.