Allow loading of configuration files from configuration subdirectory.
In a move towards a cleaner code base, we are now creating a specific directory for configuration files (of which there can be many). This makes it easier to modularise the code, as well as reduces the temptation to have the root writable for where the installer might write an initial config.

Closes #2058