Replacing warmup with an installer object
This takes the warmup code, separates the action code from the template, and creates web installer off of a central installer class. The root class contains utility methods common to all installers.

This was done to make it possible to write other installers, for example scriptable CLI installers, which utilise a common set of code.

Closes #2060