Reverting hard coded config id load in Config class.
Commit d9f3305c1cd5dea3de11636fa353dc351a505209 added a check for a specific hard coded config ID on load. This is not compatible
with Mongo (and similar NoSQL databases) which require BSON keys. 

I believe the purpose of this fix was to work around the multiple config object bug. This was fixed, but existing config objects still generate a warning in diagnostics and test. I believe the correct solution for this is a separate tidy up tool to delete the older objects.