Using safer PDO connection and statement installer.
Only the MySQL PDO library can execute multiple statements at once, and it does so badly. This is an artefact of the mysql library, and not a supported PDO feature. 

Because it is likely to be "fixed" in future, and because other databases don't support this "feature", we should only execute single statements at once.