Apache config installer now merges files.
Previously, when the installer has write access over the .htaccess file (which only really ever happens with the CLI installer), the new config was appended onto the old. Usually this is fine, but if you re-run an install (e.g. after a failed install) or run over an old Known install, this would cause problems.

Solution was to use a more sophisticated copy method, together with markers in file, to cut and paste the new config into the old.