Use queue management to do the dispatching, rather than calling an API endpoint
While a service endpoint is cool, it is problematic in real life as stuff that takes a while and uses a lot of resources can
present a DoS risk. 

The old method of dispatching within a CLI service was good, but maintaining an open connection to the database presented its own problems. 

Performing a dispatch using a separate process, via a system call to a separate Known CLI command, should be more reliable.