Disable event dispatch API service
For now, and until long running and intensive tasks can be safely dispatched via the web service, lets disable the dispatch endpoint.

Dispatching should only be done via the CLI service, although the list management endpoints are still useful as it removes the need for the service thread to maintain a connection to the database.