Adding HSTS support to Webservice
HTTP Strict Security Policy is a method by which a web server can instruct a client to address any future requests to a secure endpoint to always use the secure endpoint.

Previously, if a Known webservice call was addressed at http://example.com but was forwarded to https://example.com, future requests would also follow the Location headers. 

Now, if HSTS headers are set on https://example.com, and are found to be valid, any future request to http://example.com will automatically be rewritten to call the secure endpoint.

Refs #2181