Only display notifications for entities that are available.
It was reported that notifications could throw an exception in some cases. This happened when the notifications endpoint (which uses the email template to render the message) attempted to do so on a target that was no longer available. This meant that notifications sent on posts which were then deleted would cause all notifications to become unavailable.

Solution was to check that the target was available before it was rendered. 

Fixes #2262