Adding fallbacks for common variable names
Some MicroPub clients seem to use different variable names, and this causes some unexpected behaviour when it comes to what we accept. We should be gracious in what we accept, so lets look for "tags" if "categories" isn't found, and do similar things for "name" => "title" and allow a description field for bookmarks.

Closes #1994