Security: Adding download size limit for webservice calls.
Previously, receiving large content from webservice calls (e.g. by linking to a large file, which triggers a urlunfurl and webmention) could cause PHP to run out of memory.

Solution was to impose a file size limit for content retrieved (with a generous default of 10mb). This patch also introduces a new config option $config->webservice_max_download, which is the size in bytes.

Closes #2292