No longer throw an exception if auth token is not found
Various micropub clients were having problems authenticating if the OAuth2 plugin was activated.

This was because the OAuth2 plugin assumed that all authentication tokens were for it, but this was not the case. 

Solution was to no longer throw an exception if a token wasn't found in storage, but rather, just silently log a debug message.

Closes #2326