Moving MongoDB over to composer
Using composer to install mongodb support via composer, so we can finally get rid of /external.

Note, this is using a forked version of the mongodb client libraries, based on those which were installed at the time. The reason being is that no tagged version of the client appear to be compatible. Todo, would be to modify the code to use the later versions, however I don't currently have the time, and the mongo drivers are deprecated in any case.