Adding a more flexible url page shell override mechanism
Lets do this a little bit better, and allow a mechanism for specifying a page shell for a url path.

This lets you specify everything under admin/ and settings/ to have its own path, but leaves the door open for other url roots to have their own paths as well.